+++ MapsIndoors.framework/Headers/MPBeaconCollection.h
@@ -1,0 +1,16 @@
+//
+//  MPBeaconCollection.h
+//  MapsIndoors
+//
+//  Created by Daniel Nielsen on 26/02/15.
+//  Copyright (c) 2015 MapsPeople A/S. All rights reserved.
+//
+
+#import "MPBeacon.h"
+
+
+@interface MPBeaconCollection : MPJSONModel
+
+@property NSArray<MPBeacon*><MPBeacon>* list;
+
+@end