+++ MapsIndoors.framework/Headers/MPDistanceMatrixRows.h
@@ -1,0 +1,12 @@
+#import <Foundation/Foundation.h> 
+#import "MPJSONModel.h"
+#import "MPDistanceMatrixElements.h"
+
+
+@protocol MPDistanceMatrixRows
+@end
+
+
+@interface MPDistanceMatrixRows : MPJSONModel
+	@property (strong) NSArray<MPDistanceMatrixElements*><MPDistanceMatrixElements>* elements;
+@end